% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_COMPARISON.R
\name{get_bayes_factor_hm}
\alias{get_bayes_factor_hm}
\title{Bayes Factor via Harmonic mean}
\usage{
get_bayes_factor_hm(harmonic_mean1, harmonic_mean2)
}
\arguments{
\item{harmonic_mean1}{harmonic mean obtained from model 1}

\item{harmonic_mean2}{harmonic mean obtained from model 2}
}
\value{
\code{"bayes_factor"} bayes factor of two models up for comparison. Obtained via their harmonic means
}
\description{
Bayes Factor of two competing models obtained via the ratio of their two harmonic means
}
\examples{

bayes_factor = get_bayes_factor_hm(harmonic_mean1, harmonic_mean2)


}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
