% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_COMPARISON.R
\name{get_harmonic_mean}
\alias{get_harmonic_mean}
\title{Harmonic mean of Log likelihood}
\usage{
get_harmonic_mean(log_like_vec)
}
\arguments{
\item{log_like_vec}{Log Likelihood at each iteration of the MCMC algorithm}
}
\value{
\code{"harmonic_mean"}  harmonic mean of log likelihood vector
}
\description{
Harmonic mean of Log likelihood. Used in conjunction with \code{"get_bayes_factor_harmonic_means"} to get the Bayes Factor between two competing models
}
\examples{

harmonic_mean = get_harmonic_mean(log_like_vec)


}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
