% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_CRITICISM.R
\name{RUN_MCMC_REPS}
\alias{RUN_MCMC_REPS}
\title{Run MCMC sampler for a number of reps to facilitate Model Criticism}
\usage{
RUN_MCMC_REPS(
  epidemic_data,
  root_folder,
  model_type = list(FLAG_BASE = TRUE, FLAG_SSE = FALSE, FLAG_SSI = FALSE),
  modelling_specs = modelling_specs
)
}
\arguments{
\item{epidemic_data}{data from the epidemic, namely daily infection counts}

\item{root_folder}{root folder location in which to store the mcmc results for \code{"n_reps"}}

\item{modelling_specs}{A list of specifications;
\itemize{
  \item \code{"n_reps"} - Number of repetitions of the mcmc sampler. Multiple runs required for model criticism
  \item \code{"n_mcmc"} - Number of iterations of the mcmc sampler }}
}
\value{
saves the \code{"mcmc_output"} int the \code{"root_foler"}/\code{"rep"} location
}
\description{
MCMC algorithm with Adaptation for obtaining samples from the parameters of a
super-spreading events (SSE) epidemic model
}
\examples{

RUN_MCMC_REPS(epidemic_data, results_folder)

}
\author{
Hannah Craddock, Xavier Didelot, Simon Spencer
}
